+++ app/build.gradle
@@ -10,8 +10,8 @@ android {
         applicationId "knack.weather"
         minSdkVersion 14
         targetSdkVersion 24
-        versionCode 4
-        versionName "1.3.0(test-build)-2"
+        versionCode 5
+        versionName "1.3.0(test-build)-3"
     }
     buildTypes {
         release {
