+++ app/src/main/res/layout/activity_weather.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
+<FrameLayout
+    android:id="@+id/FragmentContainer"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -10,58 +11,4 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context=".WeatherActivity"
     android:orientation="vertical">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:id="@+id/linearLayout">
-        <android.support.design.widget.TextInputLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_weight="1">
-            <EditText
-                android:id="@+id/CityEditText"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/label.explain.your_city"
-                />
-        </android.support.design.widget.TextInputLayout>
-        <Button
-            android:id="@+id/RefreshButton"
-            android:layout_width="42dp"
-            android:layout_height="42dp"
-            android:layout_weight="0"
-            android:background="@mipmap/ic_refresh_black_24dp"
-            />
-    </LinearLayout>
-    <Button
-        android:id="@+id/GetMyWeatherButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="@string/label.widget_text.get_weather"
-        android:layout_below="@+id/linearLayout"
-        android:layout_centerHorizontal="true"
-        />
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/LocationFloatingButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        android:src="@mipmap/ic_location_on_white_48dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true"
-        />
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/MicFloatingButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:src="@mipmap/ic_mic_white_24dp"
-        android:layout_alignTop="@+id/LocationFloatingButton"
-        android:layout_centerHorizontal="true"
-        />
-
-</RelativeLayout>
+</FrameLayout>
