+++ app/src/main/res/values-ru/string.xml
@@ -122,7 +122,9 @@
     <!-- Уведомления -->
     <string name="message.notify.city_is_remove">Город удалён</string>
     <string name="message.action.undo">ОТМЕНИТЬ</string>
-    <string name="message.action.activate_gps">Активируйте геолокацию!</string>
+    <string name="message.action.activate_gps">
+        Не могу получить ваше местоположение, возможно у вас не активирована геолокация!
+    </string>
     <string name="message.notify.permission">
         Местоположение небоходимо для поиска города, в котором вы находитесь
     </string>
